+++ guacamole-ext/src/main/java/org/glyptodon/guacamole/net/auth/Connection.java
@@ -23,8 +23,8 @@
 package org.glyptodon.guacamole.net.auth;
 
 import java.util.List;
+import java.util.Map;
 import org.glyptodon.guacamole.GuacamoleException;
-import org.glyptodon.guacamole.net.auth.attribute.ObjectAttributeSet;
 import org.glyptodon.guacamole.protocol.GuacamoleConfiguration;
 
 /**
@@ -86,6 +86,25 @@ public interface Connection extends Identifiable, Connectable {
     public void setConfiguration(GuacamoleConfiguration config);
 
     /**
+     * Returns all attributes associated with this connection.
+     *
+     * @return
+     *     A map of all attribute identifiers to their corresponding values,
+     *     for all attributes associated with this connection.
+     */
+    Map<String, String> getAttributes();
+
+    /**
+     * Replaces all attributes associated with this connection with the
+     * attributes in the given map.
+     *
+     * @param attributes
+     *     A map of all attribute identifiers to their corresponding values,
+     *     for all attributes associated with this connection.
+     */
+    void setAttributes(Map<String, String> attributes);
+
+    /**
      * Returns a list of ConnectionRecords representing the usage history
      * of this Connection, including any active users. ConnectionRecords
      * in this list will be sorted in descending order of end time (active
@@ -100,18 +119,5 @@ public interface Connection extends Identifiable, Connectable {
      *                            denied.
      */
     public List<? extends ConnectionRecord> getHistory() throws GuacamoleException;
-
-    /**
-     * Returns all attributes associated with this connection.
-     *
-     * @return
-     *     An ObjectAttributeSet of all attributes associated with this
-     *     connection
-     *
-     * @throws GuacamoleException
-     *     If an error occurs while retrieving the attributes, or if reading
-     *     attributes is not allowed.
-     */
-    ObjectAttributeSet getAttributes() throws GuacamoleException;
 
 }
