+++ guacamole-ext/src/main/java/org/glyptodon/guacamole/net/auth/User.java
@@ -22,8 +22,8 @@
 
 package org.glyptodon.guacamole.net.auth;
 
+import java.util.Map;
 import org.glyptodon.guacamole.GuacamoleException;
-import org.glyptodon.guacamole.net.auth.attribute.ObjectAttributeSet;
 import org.glyptodon.guacamole.net.auth.permission.ObjectPermissionSet;
 import org.glyptodon.guacamole.net.auth.permission.SystemPermissionSet;
 
@@ -51,6 +51,25 @@ public interface User extends Identifiable {
      * @param password The password to set.
      */
     public void setPassword(String password);
+
+    /**
+     * Returns all attributes associated with this user.
+     *
+     * @return
+     *     A map of all attribute identifiers to their corresponding values,
+     *     for all attributes associated with this user.
+     */
+    Map<String, String> getAttributes();
+
+    /**
+     * Replaces all attributes associated with this user with the attributes in
+     * the given map.
+     *
+     * @param attributes
+     *     A map of all attribute identifiers to their corresponding values,
+     *     for all attributes associated with this user.
+     */
+    void setAttributes(Map<String, String> attributes);
 
     /**
      * Returns all system-level permissions given to this user.
@@ -119,17 +138,5 @@ public interface User extends Identifiable {
      *     permissions is not allowed.
      */
     ObjectPermissionSet getUserPermissions() throws GuacamoleException;
-
-    /**
-     * Returns all attributes associated with this user.
-     *
-     * @return
-     *     An ObjectAttributeSet of all attributes associated with this user.
-     *
-     * @throws GuacamoleException
-     *     If an error occurs while retrieving the attributes, or if reading
-     *     attributes is not allowed.
-     */
-    ObjectAttributeSet getAttributes() throws GuacamoleException;
 
 }
